refactor: Header and dependency refactor.
