+++ Editor/Source/Ludus/Editor/Commands/Requests/Panels.cpp
@@ -1,0 +1,20 @@
+#include "pch.h"
+
+#include <memory>
+
+#include <Ludus/Editor/Commands/Requests/Panels.h>
+#include <Ludus/Editor/Panels/ViewportPanel.h>
+
+namespace Ludus::Editor::Commands::Requests::Panels
+{
+	void AddViewport(const RequestCommand::AddViewport& command, CommandContext& context)
+	{
+		context.Panels.Register(std::make_unique<Ludus::Editor::Panels::ViewportPanel>());
+	}
+
+	void SetExecutionMode(const RequestCommand::SetExecutionMode& command, CommandContext& context)
+	{
+		context.EditorContext.State.ExecutionMode = command.Mode;
+		context.EditorContext.State.ExecutionManager.Apply(context.SystemContext, command.Mode);
+	}
+}
