+++ Engine/Include/Ludus/Engine/Events/EventBus.h
@@ -3,24 +3,28 @@
 #include <unordered_map>
 #include <vector>
 
-#include <Ludus/Engine/Events/Event.h>
-#include <Ludus/Engine/Events/Eventhandler.h>
 #include <Ludus/Engine/Events/EventType.h>
+
+namespace Ludus::Engine::Events
+{
+	struct Event;
+	struct EventHandler;
+}
 
 namespace Ludus::Engine::Events
 {
 	class EventBus
 	{
 	private:
-		std::unordered_map<EventType, std::vector<Eventhandler*>> m_EventHandlers;
+		std::unordered_map<EventType, std::vector<EventHandler*>> m_EventHandlers;
 		std::unordered_map<EventType, std::vector<const Event*>> m_EventQueue;
 
 	public:
 		EventBus() = default;
 		~EventBus() = default;
 
-		void Subscribe(EventType type, Eventhandler& handler);
-		void Unsubscribe(EventType type, Eventhandler& handler);
+		void Subscribe(EventType type, EventHandler& handler);
+		void Unsubscribe(EventType type, EventHandler& handler);
 
 		void PublishEvent(const Event& event);
